==> jupyter_notebook/notebook/02_EDA.ipynb <==
# 02. Exploratory Data Analysis
This part of the notebook is not finished yet, but meant to explore the data before running it in BGCflow

import pandas as pd
import seaborn as sns

df_bioproject = pd.read_csv("../tables/df_MAGs_bioproject.csv")
df_tax = pd.read_csv("../tables/df_MAGs_bgcflow.csv")

df = pd.merge(df_tax, df_bioproject, on="genome_id")

df.head(1)

df.columns

df_p__Nitrospirota_MAGs = df[df.loc[:, "GTDBTax"].str.contains("p__Nitrospirota")]
df_p__Nitrospirota_MAGs.shape

df_p__Nitrospirota_MAGs.genus.value_counts().plot(kind="barh")

df_p__Nitrospirota_all = pd.read_csv("../tables/p__Nitrospirota_all.csv")

df_p__Nitrospirota_all.genus.value_counts().plot(kind="barh")
